requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mssql-django==1.4
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pyodbc==5.1.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
